requirements.txt
pillow
numpy
matplotlib
opencv-contrib-python
scikit-image
scipy
toml